+++ ChangeLog.md
@@ -4,3 +4,9 @@
 
 -  Fixes bugs on `Request`
 -  Fixes bugs on `JWT` and `payload()` function.
+
+# Change Logs [v1.2.7]
+
+## [Wed, 21-08-2024]
+
+-  Update API `route` method to allow a particular request method and a particular controller method